moved TrackStatus to enum folder
