+++ motorsports-manager-backend/business-logic/src/main/java/io/sdet/msm/business/profile/TrackRecord.java
@@ -1,5 +1,6 @@
 package io.sdet.msm.business.profile;
 
+import io.sdet.msm.enums.TrackStatus;
 import lombok.Builder;
 import lombok.Data;
